% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_results}
\alias{session_results}
\title{session_results}
\usage{
session_results(session_url)
}
\arguments{
\item{session_url}{}
}
\value{
data frame, class session
}
\description{
Parse session results from webpage
}
\details{
Should add some means of trying to cleverly distinguish
results tables from other tables
could add class for the series, i.e. v8sc, f1, motogp, wrc
}
\author{
Jotham Apaloo
}
\seealso{
\code{link{readHTMLTable}}
}

