% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{sitemap}
\alias{sitemap}
\title{2014 V8Supercars Championship Sitemap}
\format{A character vector with 2031 values}
\source{
\url{http://www.XMLSitemapGenerator.org/}
}
\usage{
sitemap
}
\description{
http://www.v8supercars.com.au/championship/2014-v8-supercars-championship
}
\keyword{datasets}

