% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_results}
\alias{session_results}
\title{session_results}
\usage{
session_results(session_url)
}
\arguments{
\item{session_url}{character, link to v8supercars.com.au
results page}
}
\value{
data frame, class session
}
\description{
Parse session results from webpage
}
\author{
Jotham Apaloo
}
\seealso{
\code{link{readHTMLTable}}
}

